Fix 14 of 17 nested/Spark test failures uncovered by toArrow polyfill
Follow-up to the previous commit that introduced
``spark_dataframe_to_arrow``.  Three classes of pre-existing test
failures surfaced once the Spark→Arrow path actually started running
on PySpark 3.5 / Java 21:

- ``DataFrame.toPandas()`` calls in the nested-type tests still ran
  the broken JVM-Arrow allocator path (Java 17+ ``sun.misc.Unsafe``
  regression).  Expose ``spark_dataframe_to_pandas`` — a public alias
  for the in-module ``_to_pandas_no_arrow`` helper — and route the
  three tests through it.
- ``normalize_nested`` in the struct-test helpers expected dict-shaped
  rows (the Arrow toPandas behavior); without Arrow, PySpark hands
  struct columns back as ``Row`` objects.  Collapse via
  ``Row.asDict(recursive=True)`` so equality assertions stay direct on
  both paths.
- ``test_from_spark_type_preserves_item`` asserted that Spark
  ``LongType`` collapses to the abstract ``DataTypeId.INTEGER`` family
  tag.  The implementation correctly preserves the 64-bit width and
  returns ``INT64``; update the assertion to match.

Three failures remain, all genuine semantic issues in
``cast_spark_list_column_to_map`` / ``cast_spark_list_column``
(unioning extra null keys across rows, list cast filling) — those
need a separate investigation in the cast layer and aren't part of
this change.

+++ python/tests/test_yggdrasil/test_data/test_types/test_nested/test_array/test_spark.py
@@ -55,7 +55,10 @@ class TestSparkDtype:
         )
 
         assert isinstance(rebuilt, ArrayType)
-        assert rebuilt.item_field.dtype.type_id == DataTypeId.INTEGER
+        # Spark ``LongType`` is unambiguously 64-bit; the rebuilt dtype
+        # must preserve that width rather than collapse to the abstract
+        # ``INTEGER`` family tag.
+        assert rebuilt.item_field.dtype.type_id == DataTypeId.INT64
         assert rebuilt.item_field.nullable is False
 
     def test_handles_spark_type_only_for_array(self, spark) -> None:  # noqa: F811
